--- ttnn/cpp/ttnn/operations/data_movement/untilize/device/untilize_op.hpp
@@ -1,42 +1,0 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include <optional>
-
-#include "ttnn/tensor/tensor.hpp"
-#include "ttnn/operation.hpp"
-
-namespace ttnn::operations::data_movement {
-
-uint32_t get_pf_type(bool output_is_sharded, const Tensor& tensor);
-
-namespace untilize_helpers {
-
-uint32_t get_num_cores(CoreCoord grid_size, uint32_t nblocks);
-
-}  // namespace untilize_helpers
-
-struct Untilize {
-    const tt::tt_metal::MemoryConfig output_mem_config;
-    const bool use_multicore;
-    const bool use_pack_untilize;
-    const bool fp32_dest_acc_en;
-    const std::optional<CoreRangeSet> sub_core_grids;
-    const bool enough_space_width;
-    const bool enough_space_height;
-    const uint32_t pf_type;
-
-    void validate(const std::vector<Tensor>& input_tensors) const;
-    std::vector<ttnn::TensorSpec> compute_output_specs(const std::vector<Tensor>& input_tensors) const;
-    tt::tt_metal::operation::ProgramWithCallbacks create_program(
-        const std::vector<Tensor>& input_tensors, std::vector<Tensor>& output_tensors) const;
-    tt::tt_metal::operation::OpPerformanceModelGeneral<std::vector<Tensor>> create_op_performance_model(
-        const std::vector<Tensor>& input_tensors,
-        const std::vector<std::optional<const Tensor>>& optional_input_tensors,
-        std::vector<Tensor>& output_tensors) const;
-};
-
-}  // namespace ttnn::operations::data_movement
